Make LonelyTimer depend on the live package instead of the local version
The `../../` local dependency is apparently causing issues with Swift Package Index builds.
The local version should still be used by Xcode when using the `ComposableIdentifier` workspace, so I'll make `LonelyTimer` depend on the online version to check if it solves the issue.
Otherwise, I'll need to check if SPI provides an API to build only the top-level package.